+++ services/backend/src/app.module.ts
@@ -1,7 +1,8 @@
-import { MiddlewareConsumer, Module, NestModule, RequestMethod } from '@nestjs/common';
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { AppController } from './app.controller.js';
 import { AppService } from './app.service.js';
 import { AuthModule } from './auth/auth.module.js';
+import { AuthController } from './auth/auth.controller.js';
 import { FirebaseAuthenticationMiddleware } from './auth/authentication/firebase-authentication.middleware.js';
 
 @Module({
@@ -11,9 +12,6 @@ import { FirebaseAuthenticationMiddleware } from './auth/authentication/firebase
 })
 export class AppModule implements NestModule {
   configure(consumer: MiddlewareConsumer): void {
-    consumer.apply(FirebaseAuthenticationMiddleware).forRoutes({
-      path: '',
-      method: RequestMethod.GET
-    });
+    consumer.apply(FirebaseAuthenticationMiddleware).forRoutes(AuthController);
   }
 }